Fix the definition of mag_{}_cModel for HDF5 DC2ObjectCatalog
This bug was introduced in cc8bbc14 as part of #369.
There was a bug where mag_{}_cmodel was defined baesd on mag_{}_cModel

This should have only affected Run 1.1p and HSC.
because it was triggered if the modelfit mags were not already calculated
and only for the default (HDF5) catalogs.  The Parquet class was fine.